Move store contact to Dashboard
