+++ sphinx/screens/dashboard/dashboard/build.gradle
@@ -42,6 +42,7 @@ dependencies {
     api project(path: ':sphinx:application:data:concepts:repositories:concept-repository-feed')
     api project(path: ':sphinx:application:data:concepts:repositories:concept-repository-media')
     api project(path: ':sphinx:application:common:menus:menu-bottom-scanner')
+    api project(path: ':sphinx:application:common:wrappers:wrapper-mqtt')
 
     api project(path: ':sphinx:application:common:wrappers:wrapper-chat')
